Support JSON or HTML responses depending on the incoming request (#2).
The request object is checked to see if it expects a JSON response. Then an appropriate instance of ResponseBuilder is used to build the response.

This is not yet completed, as validation errors are not returned in the response (related to #1) and redirects are not yet supported, which is a common scenario in HTML responses.